отрисовка галлереи с одновременным добавлением обработчика на контейнер
