+++ js/open-picture.js
@@ -1,52 +1,29 @@
 import { isEscapeKey, isEnterKey } from './util.js'; //1. почему недостаточно импортировать только в main.js?
-import { thumbnailsList } from './thumbnail.js';
 
 const body = document.querySelector('body');
 const bigPicture = document.querySelector('.big-picture'); //окно большой картинки
 const pictureCancelButton = bigPicture.querySelector('.big-picture__cancel'); //кнопка закрытия окна с картинкой
 const socialCommentCount = bigPicture.querySelector('.social__comment-count');
 const commentsLoader = bigPicture.querySelector('.comments-loader');
-const bigPicturImg = bigPicture.querySelector('img');
-
-
-
-
 
 //закрытие картинки нажатием ESCAPE на DOCUMENT
 const onDocumentKeydown = (evt) => {
   if (isEscapeKey(evt)) {
     evt.preventDefault();
     bigPicture.classList.add('hidden');
+    body.classList.remove('modal-open');
   }
 };
 
-
 //функция открытия картинки
-const openPicture = () => {
+const openPicture = (data) => {
   bigPicture.classList.remove('hidden');
   document.addEventListener('keydown', onDocumentKeydown);
   body.classList.add('modal-open');
   socialCommentCount.classList.add('hidden');
   commentsLoader.classList.add('hidden');
+
 };
-
-
-//ОБРАБОТЧИК открытие большой картинки нажатием клавиши ENTER
-thumbnailsList.addEventListener('keydown', (evt) => {
-  if (isEnterKey(evt)) {
-    if (evt.target.classList.contains('picture')) {
-      openPicture();
-    }
-  }
-});
-
-//ОБРАБОТЧИК открытие большой картинки нажатием мышки
-thumbnailsList.addEventListener('click', (evt) => {
-  if (evt.target.classList.contains('picture__img')) {
-    openPicture();
-  }
-});
-
 
 //функция зактытия картинки
 const closePicture = () => {
@@ -66,10 +43,6 @@ pictureCancelButton.addEventListener('keydown', (event) => {
     closePicture();
   }
 });
+//
 
-/*
-thumbnailsList.addEventListener('click', () => {
-  bigPicturImg.src = 'cdzc';
-});
-*/
-export {bigPicture, bigPicturImg};
+export {bigPicture, openPicture, closePicture};
